requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.0
python-dotenv>=1.0
openai>=1.0
requests>=2.31

# Optional for real LLM optimization/token counting:
transformers>=4.40
langchain-openai>=0.1
litellm>=1.0

litellm>=1.82.6
dspy-ai>=3.2.0
rouge>=1.0.1
git+https://github.com/SalesforceAIResearch/promptomatix.git